Update project version to 0.6.15 and add server components
- Bump project version from 0.6.14 to 0.6.15.
- Introduce new server components: Config, Socket, Workers, and Status for enhanced server functionality.
- Implement configuration loading, socket setup, worker thread management, and system status reporting.
- Update CMakeLists.txt to include new server source files for compilation.
- Enhance documentation in Geruest.cpp to reflect the new server architecture and methods.
